refactor: 테스트용 Dao 객체를 test 패키지로 수정
