+++ src/test/java/wooteco/subway/Infrastructure/line/MemoryLineDao.java
@@ -1,5 +1,6 @@
 package wooteco.subway.Infrastructure.line;
 
+import wooteco.subway.Infrastructure.line.LineDao;
 import wooteco.subway.domain.line.Line;
 
 import java.util.*;
